v0.0.2 (#8)
* updated travis.yml

* Updated travis.yml

* fixed menu toggle. closes #4

* improved spinner centering. closes #7

* improved project view layouts. closes #5

* implemented basic e2e tests

* added e2e tests to travis

* updated travis.yml

* reverted change to travis.yml

* updated travis.yml

* updated travis.yml

* reverted updated to travis.yml

* fixed travis config

* update travis.yml

* added experience module closes #6

* update travis.yml

* fixed build errors

* fixed failing tests

* fixed deploy.sh

* converted to sass, closes #9 changed deploy.sh fixes #10

+++ deploy.sh
@@ -2,16 +2,18 @@
 # See https://medium.com/@nthgergo/publishing-gh-pages-with-travis-ci-53a8270e87db
 
 if [[ $TRAVIS_BRANCH == 'master' ]]; then
-    set -o errexit
+    if [[ $TRAVIS_PULL_REQUEST == "false" ]]; then
+        set -o errexit
 
-    # config
-    git config --global user.email "george.allen93@outlook.com"
-    git config --global user.name "GeorgeA93"
+        # config
+        git config --global user.email "george.allen93@outlook.com"
+        git config --global user.name "GeorgeA93"
 
-    # deploy
-    cd dist
-    git init
-    git add .
-    git commit -m "Deploy to Github Pages"
-    git push --force --quiet "https://${GH_TOKEN}@github.com/${GITHUB_REPO}.git" master:gh-pages > /dev/null 2>&1
+        # deploy
+        cd dist
+        git init
+        git add .
+        git commit -m "Deploy to Github Pages"
+        git push --force --quiet "https://${GH_TOKEN}@github.com/${GITHUB_REPO}.git" master:gh-pages > /dev/null 2>&1
+    fi
 fi
